pyproject.toml
[project]
name = "mcp-server-demo"
version = "0.1.0"
description = "Sample MCP server PoC for testing (BITS mockup)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.52.0",
    "azure-identity>=1.23.0",
    "fastmcp>=2.3.0",
    "mcp[cli]>=1.8.0",
    "openai>=1.81.0",
    "pandas>=2.2.3",
    "pymssql>=2.3.4",
    "ruff>=0.11.8",
]

[tool.uv]
no-binary-package = ["pymssql"]

[tool.setuptools.packages.find]
include = ["auth", "business_request"]
